Overhaul of function and effectiveness
It dawned on me that a better approach would be to (try) to
unifify both touch and hover, rather than leave them as two
seperate entities. So after some painful trial and error, I
found a way to simulate :hover with JS mouse events (was not
as straightforward as expected since mouse enter/leave do not
bubble and thus cannot be used in a delegated sense, and mouse
over/out will trigger when traversing internal elements,
requiring some nest checking). It seems to work on iOS Safari
and standard Chrome, but much testing remains.

Aside from that, nest checking has been added to the touch
event checker, so now spans in an anchor link will not break
the hover effect.

The new strategy combines touch and hover into a single class
.hovertouch, so :hover can be abandoned as it should be.